Cleanup

+++ RealEngine/Font.cpp
@@ -12,7 +12,7 @@ TTF_Font* real::Font::GetFont() const {
 
 real::Font::Font(const std::string& fullPath, unsigned int size) : m_font(nullptr), m_size(size)
 {
-	m_font = TTF_OpenFont(fullPath.c_str(), size);
+	m_font = TTF_OpenFont(fullPath.c_str(), static_cast<int>(size));
 	if (m_font == nullptr) 
 	{
 		throw std::runtime_error(std::string("Failed to load font: ") + SDL_GetError());
